requirements.txt
torch==2.5.1
jax
flax
transformers
Pillow
numpy
matplotlib